requirements.txt
pymysql>=1.1.0
redis>=5.0.0
pandas>=2.1.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
loguru>=0.7.0
requests>=2.31.0
